requirements.txt
python-dotenv==1.0.0
easyocr==1.7.1
openai==1.3.8
pylint==3.0.2